+++ clients/include/testing_spr.hpp
@@ -26,8 +26,8 @@ hipblasStatus_t testing_spr(const Arguments& argus)
     hipblasFillMode_t uplo      = char2hipblas_fill(char_uplo);
 
     int    abs_incx = incx < 0 ? -incx : incx;
-    size_t A_size   = N * (N + 1) / 2;
-    size_t x_size   = abs_incx * N;
+    size_t A_size   = size_t(N) * (N + 1) / 2;
+    size_t x_size   = abs_incx * size_t(N);
 
     // argument sanity check, quick return if input parameters are invalid before allocating invalid
     // memory
