+++ clients/include/testing_syr.hpp
@@ -27,8 +27,8 @@ hipblasStatus_t testing_syr(const Arguments& argus)
     hipblasFillMode_t uplo      = char2hipblas_fill(char_uplo);
 
     int    abs_incx = incx < 0 ? -incx : incx;
-    size_t A_size   = lda * N;
-    size_t x_size   = abs_incx * N;
+    size_t A_size   = size_t(lda) * N;
+    size_t x_size   = abs_incx * size_t(N);
 
     // argument sanity check, quick return if input parameters are invalid before allocating invalid
     // memory
